requirements.txt
flask
tensorflow==2.17.1
requests
pillow
gunicorn